Support for expand_user
Uses path.expand.  I looked at a more complicated way that would
implement the passwd(5) reading approach (on POSIX) used by the
Python module, but it's not clear that this would be more portable or
robust.  The path.expand version will (apparently) only work on some unixes
and only if readline is installed